rename: move class - MemberRoleFactory to factory package
