+++ src/main/java/co/fineants/api/domain/member/config/MemberRoleConfig.java
@@ -3,7 +3,7 @@ package co.fineants.api.domain.member.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import co.fineants.api.domain.member.service.MemberRoleFactory;
+import co.fineants.api.domain.member.domain.factory.MemberRoleFactory;
 import co.fineants.api.domain.member.service.RoleService;
 
 @Configuration
